refactor: error handling and components naming
