+++ frontend/src/app/queryClient.ts
@@ -1,14 +1,25 @@
-import { QueryClient } from "@tanstack/react-query";
+import { QueryCache, QueryClient } from "@tanstack/react-query";
 import { AxiosError } from "axios";
+import { toast } from "sonner";
 
 const queryClient = new QueryClient({
+  queryCache: new QueryCache({
+    onError: (error) =>
+      toast.error("Something went wrong", {
+        description: error?.message || "An unexpected error occurred.",
+      }),
+  }),
   defaultOptions: {
     queries: {
+      staleTime: 1000 * 60 * 2, // 2 minutes
+      refetchOnWindowFocus: false,
       retry: (failureCount, error: unknown) => {
         // Don't retry validation errors
         if (error instanceof Error && error.message.includes("Invalid data")) {
           return false;
         }
+        // TODO: Consider handling 401 errors with clearing auth state
+
         // Type guard for AxiosError
         // Retry on server errors (status 500 and above) up to 3 times
         if (error instanceof AxiosError) {
