+++ frontend/src/features/auth/components/SignUpForm.tsx
@@ -5,7 +5,6 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import z from "zod";
 import { useAuth } from "@/features/auth/hooks/useAuth";
-import { useEffect } from "react";
 import { Link } from "react-router";
 import { SignUpRequestSchema } from "@/features/auth/schemas";
 import BrandButton from "@/shared/components/BrandButton";
@@ -13,12 +12,11 @@ import { CatDocLogo } from "@/shared/components/ui/CatDocLogo";
 import { Separator } from "@/shared/components/ui/Separator";
 import { Muted } from "@/shared/components/ui/typography/Paragraph";
 import { motion } from "framer-motion";
-import { toast } from "sonner";
 
 type Form = z.infer<typeof SignUpRequestSchema>;
 
 const SignUpForm = () => {
-  const { signUp, isLoading, isSignUpError, signUpError } = useAuth();
+  const { signUp, isLoading } = useAuth();
 
   const form = useForm<Form>({
     resolver: zodResolver(SignUpRequestSchema),
@@ -31,14 +29,6 @@ const SignUpForm = () => {
       password: "",
     },
   });
-
-  useEffect(() => {
-    if (isSignUpError) {
-      toast.error("Sign up failed", {
-        description: signUpError?.message || "An error occurred during registration. Please try again.",
-      });
-    }
-  }, [isSignUpError, signUpError]);
 
   const onSubmit = (data: Form) => {
     signUp(data);
